Moved loaders into its own folder and namespace
