+++ Stubble/Classes/Registry.cs
@@ -5,6 +5,7 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reflection;
 using System.Text.RegularExpressions;
+using Stubble.Core.Classes.Loaders;
 using Stubble.Core.Classes.Tokens;
 using Stubble.Core.Helpers;
 using Stubble.Core.Interfaces;
